A new function to convert any string, cell of strings or structure whose terminal nodes are strings or cell of strings to a cell of string.
A new function to build the list of dependencies of a pipeline. For each job A, a structure describes all the jobs B 
that need to be completed before job A can be submitted. A list of the exact files that job A is waiting from job B is 
also generated.