I did not have internet access for more than one week, so this is a rather big commit. Mainly, the update feature (milestone 0.7) has been implemented. The way that the pipeline manager deals with the logs folder had to be largely revised for that purpose.
Now, all modifications to the log folder are made through PSOM_PIPELINE_INIT prior to any call to 
PSOM_PIPELINE_PROCESS. This last function now does not modify the existing log, and will thus not be able 
to handle any type of failure or restart. As a consequence, the only real command that is worth using 
outside of development purposes is NIAK_RUN_PIPELINE, which is mixing the initialization, the processing, 
and the visualization.

The modifications have only been superficially tested. More tests are required before version 0.7 can be 
released.