2 major changes in this commit :
1. In the initialization of the log folder, the status/logs/job files are saved in one shot, rather than 
saving each job independently as was done before. That should speed up the process a lot.

2. The job status are now stored in the same form as the logs and job description : one variable per job, 
which is a string. That comes with two benefits : the pipeline manager should now able to deal with 
restarts where the number of jobs have changed. It is also faster to update the status of the jobs on the 
disk, because before all the status had to be saved every time.

These changes have not been tested yet.