Many changes to ensure compatibility with Octave. Session and batch modes have been tested on Octave 3.0.4. Qsub mode still needs to be tested.
The only modification to the general design is that now 
the pipeline manager is not waiting for batch/qsub to 
answer back before submitting new jobs. This is speeding
 up the submission process a lot, so PSOM should now be able to deal 
with hundreds of computing cores. It is still possible to get the errors 
in batch/qsub submission by turning on OPT.FLAG_DEBUG in 
PSOM_RUN_PIPELINE.