bug fix for the last commit.
Now the old status are loaded and updated based on left-over tags.

Some restart tags are then generated based on :
  * a 'failed' status (that's the new part)
  * a change in the job description
  * a manual user-defined tag

The children of restart jobs are set to restart

The parent of restart jobs that produce an input of a restart job are 
set to restart.

This has been tested superficially and apparently works. There should be 
more in-depths tests before release.